OCRViewFromMenu/requirements.txt
altgraph==0.17.2
black==22.8.0
cachetools==5.2.0
certifi==2022.9.14
chardet==4.0.0
charset-normalizer==2.1.1
colorama==0.4.5
customtkinter==4.6.3
cx-Logging==3.0
cycler==0.11.0
darkdetect==0.7.1
fonttools==4.37.3
future==0.18.2
google-api-core==2.10.1
google-auth==2.6.0
google-cloud-core==2.3.2
google-cloud-vision==3.1.2
googleapis-common-protos==1.56.4
grpcio==1.48.1
grpcio-status==1.48.1
idna==3.4
importlib-metadata==4.12.0
kiwisolver==1.4.4
lief==0.12.1
matplotlib==3.5.3
mccabe==0.7.0
mojimoji==0.0.12
mypy-extensions==0.4.3
numexpr==2.8.3
numpy==1.21.4
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
pandastable==0.13.0
pathspec==0.10.1
pdf2image==1.16.0
pefile==2022.5.30
Pillow==9.0.0
platformdirs==2.5.2
proto-plus==1.22.1
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyflakes==2.5.0
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyocr==0.8.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pywin32-ctypes==0.2.0
requests==2.28.1
rsa==4.9
screeninfo==0.8.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
typed-ast==1.5.4
typing_extensions==4.3.0
urllib3==1.26.12
xlrd==2.0.1
zipp==3.8.1